second commit: changed into integer

+++ print_difference.py
@@ -1,8 +1,9 @@
 #6. Create a program that ask user to input 2 numbers. Print the difference of the two numbers.
 
-num1 = float(input("Please enter a number: "))
-num2 = float(input("Please enter another number: "))
+num1 = int(input("Please enter a number: "))
+num2 = int(input("Please enter another number: "))
 
 difference = num1 - num2
 
 print(f"the difference of {num1} and {num2} is {difference}. mhuehehwahha skibidi bomni")
+
